+++ app/Http/Requests/RoomTypeRequest.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Requests;
 
+use App\Rules\Color;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 
@@ -12,7 +13,7 @@ class RoomTypeRequest extends FormRequest
         $rules = [
             'description'   => ['required', 'string', 'max:255', Rule::unique('room_types', 'description')],
             'short'         => ['required', 'string', 'max:2', Rule::unique('room_types', 'short')],
-            'color'         => ['required', 'string','regex:/^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/']
+            'color'         => ['required', 'string', new Color()]
         ];
 
         if ($this->roomType) {