+++ resources/lang/en/validation.php
@@ -142,12 +142,20 @@ return [
         'room' => [
             'already_member' => 'The user is already member of the room.'
         ],
-        'password' => 'The password must contain at least one character from each of the following four categories:: Uppercase letter (A - Z), lowercase letter (a - z), number (0 - 9), non-alphanumeric character (for example: !, $, #, or %).',
+        'password' => 'The password must contain at least one character from each of the following four categories: Uppercase letter (A - Z), lowercase letter (a - z), number (0 - 9), non-alphanumeric character (for example: !, $, #, or %).',
         'replacement_room_type' => [
             'required' => 'Replacement room type required! Rooms are still assigned to this room type.',
             'exists' => 'Replacement room type invalid! A replacement room type is required because rooms are still assigned to this room type.',
             'not_in' => 'Replacement room type invalid! A replacement room type is required because rooms are still assigned to this room type.',
-        ]
+        ],
+        'banner' => [
+            'array' => 'The message banner settings are missing!',
+            'required' => 'The message banner settings are missing!',
+        ],
+        'banner.icon' => [
+            'regex' => 'The icon css class should be in the following format: `fas fa-camera`. The class `fas` depends on the selected type and can also be for example `fab`.'
+        ],
+        'color' => ':attribute must be a color specified as a hexadecimal number (z. B. #fff oder #ffffff)!'
     ],
 
     /*
@@ -221,6 +229,16 @@ return [
         'firstname'                         => 'Firstname',
         'lastname'                          => 'Lastname',
         'user_locale'                       => 'Language',
+        'banner.enabled'                    => 'Show',
+        'banner.message'                    => 'Message',
+        'banner.color'                      => 'Text color of the banner',
+        'banner.background'                 => 'Background color of the banner',
+        'banner.title'                      => 'Title',
+        'banner.link'                       => 'Link to show after the message',
+        'banner.icon'                       => 'Icon',
+        'banner.link_text'                  => 'Link text',
+        'banner.link_target'                => 'Link target',
+        'banner.link_style'                 => 'Link style',
     ],
 
 ];