+++ tests/Feature/api/v1/SettingsTest.php
@@ -36,6 +36,15 @@ class SettingsTest extends TestCase
         setting(['pagination_page_size' => '123']);
         setting(['own_rooms_pagination_page_size' => '123']);
         setting(['room_limit' => '-1']);
+        setting(['banner' => [
+            'enabled'    => true,
+            'message'    => 'Welcome to Test!',
+            'title'      => 'Welcome',
+            'color'      => '#fff',
+            'background' => '#4a5c66',
+            'link'       => 'http://localhost',
+            'icon'       => 'fas fa-door-open',
+        ]]);
 
         $this->getJson(route('api.v1.application'))
             ->assertJson([
@@ -44,9 +53,173 @@ class SettingsTest extends TestCase
                     'pagination_page_size'           => '123',
                     'own_rooms_pagination_page_size' => '123',
                     'room_limit'                     => '-1',
+                    'banner'                         => [
+                        'enabled'    => true,
+                        'message'    => 'Welcome to Test!',
+                        'title'      => 'Welcome',
+                        'color'      => '#fff',
+                        'background' => '#4a5c66',
+                        'link'       => 'http://localhost',
+                        'icon'       => 'fas fa-door-open',
+                    ]
                 ]
             ])
             ->assertSuccessful();
+
+        setting(['banner' => [
+            'enabled'    => false,
+            'message'    => 'Welcome to Test!',
+            'title'      => 'Welcome',
+            'color'      => '#fff',
+            'background' => '#4a5c66',
+            'link'       => 'http://localhost',
+            'icon'       => 'fas fa-door-open',
+        ]]);
+
+        $this->getJson(route('api.v1.application'))
+            ->assertJson([
+                'data' => [
+                    'logo'                           => 'testlogo.svg',
+                    'pagination_page_size'           => '123',
+                    'own_rooms_pagination_page_size' => '123',
+                    'room_limit'                     => '-1',
+                    'banner'                         => [
+                        'enabled'    => false
+                    ]
+                ]
+            ])
+            ->assertSuccessful();
+    }
+
+    /**
+     * Tests that the correct application wide settings provided
+     * when listing all settings for edition.
+     *
+     * @return void
+     */
+    public function testAllApplicationSettings()
+    {
+        setting(['logo' => 'testlogo.svg']);
+        setting(['pagination_page_size' => '123']);
+        setting(['own_rooms_pagination_page_size' => '123']);
+        setting(['room_limit' => '-1']);
+        setting(['banner' => [
+            'enabled'    => true,
+            'message'    => 'Welcome to Test!',
+            'title'      => 'Welcome',
+            'color'      => '#fff',
+            'background' => '#4a5c66',
+            'link'       => 'http://localhost',
+            'icon'       => 'fas fa-door-open',
+        ]]);
+
+        $this->getJson(route('api.v1.application.complete'))->assertUnauthorized();
+        $this->actingAs($this->user)->getJson(route('api.v1.application.complete'))->assertForbidden();
+
+        $role       = factory(Role::class)->create();
+        $permission = factory(Permission::class)->create(['name' => 'settings.viewAny']);
+        $role->permissions()->attach($permission);
+        $this->user->roles()->attach($role);
+
+        $this->getJson(route('api.v1.application.complete'))
+            ->assertJson([
+                'data' => [
+                    'logo'                           => 'testlogo.svg',
+                    'pagination_page_size'           => '123',
+                    'own_rooms_pagination_page_size' => '123',
+                    'room_limit'                     => '-1',
+                    'banner'                         => [
+                        'enabled'    => true,
+                        'message'    => 'Welcome to Test!',
+                        'title'      => 'Welcome',
+                        'color'      => '#fff',
+                        'background' => '#4a5c66',
+                        'link'       => 'http://localhost',
+                        'icon'       => 'fas fa-door-open',
+                    ]
+                ]
+            ])
+            ->assertSuccessful();
+
+        setting(['banner' => [
+            'enabled'    => false,
+            'message'    => 'Welcome to Test!',
+            'title'      => 'Welcome',
+            'color'      => '#fff',
+            'background' => '#4a5c66',
+            'link'       => 'http://localhost',
+            'icon'       => 'fas fa-door-open',
+        ]]);
+
+        $this->getJson(route('api.v1.application.complete'))
+            ->assertJson([
+                'data' => [
+                    'logo'                           => 'testlogo.svg',
+                    'pagination_page_size'           => '123',
+                    'own_rooms_pagination_page_size' => '123',
+                    'room_limit'                     => '-1',
+                    'banner'                         => [
+                        'enabled'    => false,
+                        'message'    => 'Welcome to Test!',
+                        'title'      => 'Welcome',
+                        'color'      => '#fff',
+                        'background' => '#4a5c66',
+                        'link'       => 'http://localhost',
+                        'icon'       => 'fas fa-door-open',
+                    ]
+                ]
+            ])
+            ->assertSuccessful();
+    }
+
+    public function testAllApplicationSettingsReturnedOnUpdate()
+    {
+        $payload = [
+            'name'                           => 'test',
+            'logo'                           => 'testlogo.svg',
+            'favicon'                        => 'favicon.ico',
+            'pagination_page_size'           => '10',
+            'own_rooms_pagination_page_size' => '15',
+            'room_limit'                     => '-1',
+            'banner'                         => [
+                'enabled'     => false,
+                'message'     => 'Welcome to Test!',
+                'title'       => 'Welcome',
+                'color'       => '#fff',
+                'background'  => '#4a5c66',
+                'link'        => 'http://localhost',
+                'link_target' => 'self',
+                'link_style'  => 'primary',
+                'icon'        => 'fas fa-door-open',
+            ]
+        ];
+
+        $role       = factory(Role::class)->create();
+        $permission = factory(Permission::class)->create(['name' => 'settings.update']);
+        $role->permissions()->attach($permission);
+        $this->user->roles()->attach($role);
+
+        $this->actingAs($this->user)->putJson(route('api.v1.application.update'), $payload)
+            ->assertSuccessful()
+            ->assertJson([
+                'data' => [
+                    'name'                           => 'test',
+                    'logo'                           => 'testlogo.svg',
+                    'favicon'                        => 'favicon.ico',
+                    'pagination_page_size'           => 10,
+                    'own_rooms_pagination_page_size' => 15,
+                    'room_limit'                     => -1,
+                    'banner'                         => [
+                        'enabled'    => false,
+                        'message'    => 'Welcome to Test!',
+                        'title'      => 'Welcome',
+                        'color'      => '#fff',
+                        'background' => '#4a5c66',
+                        'link'       => 'http://localhost',
+                        'icon'       => 'fas fa-door-open',
+                    ]
+                ]
+            ]);
     }
 
     /**
@@ -63,6 +236,7 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
+            'banner'                         => ['enabled' => false]
         ];
 
         // Unauthorized Test
@@ -97,6 +271,7 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
+            'banner'                         => ['enabled' => false]
         ];
 
         // Unauthorized Test
@@ -134,6 +309,7 @@ class SettingsTest extends TestCase
             'pagination_page_size'           => '10',
             'own_rooms_pagination_page_size' => '15',
             'room_limit'                     => '-1',
+            'banner'                         => ['enabled' => false]
         ];
 
         // Add necessary role and permission to user to update application settings
@@ -161,18 +337,18 @@ class SettingsTest extends TestCase
         $role->permissions()->attach($permission);
         $this->user->roles()->attach($role);
 
-        $this->actingAs($this->user)->putJson(route('api.v1.application.update'),
-            [
-                'name'                           => '',
-                'favicon'                        => '',
-                'favicon_file'                   => 'notimagefile',
-                'logo'                           => '',
-                'logo_file'                      => 'notimagefile',
-                'pagination_page_size'           => 'notnumber',
-                'own_rooms_pagination_page_size' => 'notnumber',
-                'room_limit'                     => 'notnumber',
-            ]
-        )
+        $payload = [
+            'name'                           => '',
+            'favicon'                        => '',
+            'favicon_file'                   => 'notimagefile',
+            'logo'                           => '',
+            'logo_file'                      => 'notimagefile',
+            'pagination_page_size'           => 'notnumber',
+            'own_rooms_pagination_page_size' => 'notnumber',
+            'room_limit'                     => 'notnumber'
+        ];
+
+        $this->actingAs($this->user)->putJson(route('api.v1.application.update'), $payload)
             ->assertStatus(422)
             ->assertJsonValidationErrors([
                 'name',
@@ -182,7 +358,72 @@ class SettingsTest extends TestCase
                 'logo_file',
                 'pagination_page_size',
                 'own_rooms_pagination_page_size',
-                'room_limit'
+                'room_limit',
+                'banner',
+                'banner.enabled'
+            ]);
+
+        $payload = [
+            'name'                           => 'test',
+            'favicon'                        => '/storage/image/favicon.ico',
+            'logo'                           => '/storage/image/testfile.svg',
+            'pagination_page_size'           => '10',
+            'own_rooms_pagination_page_size' => '15',
+            'room_limit'                     => '-1',
+            'banner'                         => false
+        ];
+
+        $this->putJson(route('api.v1.application.update'), $payload)
+            ->assertStatus(422)
+            ->assertJsonValidationErrors([
+                'banner',
+                'banner.enabled'
+            ]);
+
+        $payload['banner'] = [
+            'enabled' => 'foo'
+        ];
+
+        $this->putJson(route('api.v1.application.update'), $payload)
+            ->assertStatus(422)
+            ->assertJsonValidationErrors([
+                'banner.enabled'
+            ]);
+
+        $payload['banner'] = [
+            'enabled' => true
+        ];
+
+        $this->putJson(route('api.v1.application.update'), $payload)
+            ->assertStatus(422)
+            ->assertJsonValidationErrors([
+                'banner.message',
+                'banner.color',
+                'banner.background',
+            ]);
+
+        $payload['banner'] = [
+            'enabled'    => true,
+            'title'      => str_repeat('a', 256),
+            'message'    => str_repeat('a', 501),
+            'link'       => 'test',
+            'link_style' => 'test',
+            'icon'       => 'test-test',
+            'color'      => 'test-test',
+            'background' => 'test-test',
+        ];
+
+        $this->putJson(route('api.v1.application.update'), $payload)
+            ->assertStatus(422)
+            ->assertJsonValidationErrors([
+                'banner.message',
+                'banner.color',
+                'banner.background',
+                'banner.link',
+                'banner.icon',
+                'banner.title',
+                'banner.link_style',
+                'banner.link_target',
             ]);
     }
 
@@ -208,6 +449,7 @@ class SettingsTest extends TestCase
                 'pagination_page_size'           => '0',
                 'own_rooms_pagination_page_size' => '0',
                 'room_limit'                     => '-2',
+                'banner'                         => ['enabled' => false]
             ]
         )
             ->assertStatus(422)
@@ -226,6 +468,7 @@ class SettingsTest extends TestCase
                 'pagination_page_size'           => '101',
                 'own_rooms_pagination_page_size' => '26',
                 'room_limit'                     => '101',
+                'banner'                         => ['enabled' => false]
             ]
         )
             ->assertStatus(422)