+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added management of servers ([#30], [#88])
 - Added artisan command to check server and meeting status, build live and historical data ([#64], [#65])
 - Added artisan command to create a new admin user ([#81])
+- Added a application banner that can be modified throught the application settings ([#89], [#94])
 
 [#1]: https://github.com/THM-Health/PILOS/issues/1
 [#3]: https://github.com/THM-Health/PILOS/pull/3
@@ -73,5 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#81]: https://github.com/THM-Health/PILOS/pull/81
 [#87]: https://github.com/THM-Health/PILOS/pull/87
 [#88]: https://github.com/THM-Health/PILOS/pull/88
+[#89]: https://github.com/THM-Health/PILOS/issues/89
+[#94]: https://github.com/THM-Health/PILOS/pull/94
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/3c8359cdb0395546fe97aeabf1a40f93002b182c...HEAD